[RENAME/#1] 프로젝트명 오타 수정
