+++ week1/practice/BirthdayCard/app/src/main/java/com/example/birthdaycard/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.example.bithdaycard.ui.theme
+package com.example.birthdaycard.ui.theme
 
 import androidx.compose.ui.graphics.Color
