+++ week1/practice/BirthdayCard/app/src/main/java/com/example/birthdaycard/ui/theme/Shape.kt
@@ -1,4 +1,4 @@
-package com.example.bithdaycard.ui.theme
+package com.example.birthdaycard.ui.theme
 
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material.Shapes
