+++ week1/practice/BirthdayCard/app/src/main/java/com/example/birthdaycard/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.example.bithdaycard.ui.theme
+package com.example.birthdaycard.ui.theme
 
 import androidx.compose.material.Typography
 import androidx.compose.ui.text.TextStyle
