+++ week1/practice/BirthdayCard/app/src/main/res/values/strings.xml
@@ -1,3 +1,3 @@
 <resources>
-    <string name="app_name">Bithday Card</string>
+    <string name="app_name">Birthday Card</string>
 </resources>
