+++ week1/practice/BirthdayCard/app/src/main/res/values/themes.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
 
-    <style name="Theme.BithdayCard" parent="android:Theme.Material.Light.NoActionBar">
+    <style name="Theme.BirthdayCard" parent="android:Theme.Material.Light.NoActionBar">
         <item name="android:statusBarColor">@color/purple_700</item>
     </style>
 </resources>
