+++ week1/practice/BirthdayCard/settings.gradle
@@ -12,5 +12,5 @@ dependencyResolutionManagement {
         mavenCentral()
     }
 }
-rootProject.name = "Bithday Card"
+rootProject.name = "Birthday Card"
 include ':app'
